Fix usage with touch
This is a important change:
Now instead of use the events mousedown and mouseover,
use the event pressup and pressmove. These events work ok with touch.
Removed the need of detect when the mouse/touch start, just checking
the state of variable in the same pressmove event.

Also set mouseEnabled=false in the lines used to show the words,
to avoid loose events over that lines.