Workaroud a bug affecting webkit on android 4.1/4.2
We need overwrite the definition in sugar-web/graphics/css/sugar.css
This bug affect us, because we need resize the matrix canvas
according to the space available. Th eeffect is the drawing in the canvas
is copied out of the canvas, over the left section of the screen.

There are more information here:
https://code.google.com/p/android/issues/detail?id=41312
I have tested every proposed workaround, and this is the only who works.