Initial implementation of use of android back button
Use the phonegap events [1] to detect the back button.
As we need load the cordova.js file, use a empty file as a fallback
too ise in Sugar. The cordova builder will replace it with
the file needed by the platform.
A complete implementation should see what stage of the game is active
and change to the previous page.

[1] http://docs.phonegap.com/en/3.5.0/cordova_events_events.md.html