Use hitArea to avoid security errors
After adding bckground images, the mouse event produce security
errors, because createjs try to read he pixel values to know
where is the mouse. This is solved setting the hitArea property
in _every_ container where we use mouse events.

This patch also resize the wordlist canvas to avoid display a
white space at bottom, the real problem need be solved.
This is a temporary change.