Update style in the create panel
Added new css to adjust the sizes according to the dpi in the screen.
Also add a gradient in the background, with rules for compatility
with the old webkit available in the android image in the xo.
Useful documentation here [1] and [2], but even that is not completly
accurate. You can't use "bottom right" to define the points
or use percentages in the radius.

[1] https://www.webkit.org/blog/175/introducing-css-gradients/
[2] http://www.the-art-of-web.com/css/radial-gradients/