Implement buttons in the wordlist in the canvas
Instead of create a transparent div floating over the canvas,
draw the buttons in the canvas itself in the case of the app
running on android.
The svg are different , because are black instead of white.
A sound button is added accoreding to the design, but we don't
use it yet.