Update lib/sugar-web with upstream bits
Recently, we pushed into sugar-web a fix for issue of not being
able to resolve "localhost", which is to use 0.0.0.0 instead.

Therefore, it is recommended that we update web activities with
upstream sugar-web from the following commit or higher.

commit: e17b17dc36e4dbe9d215daa8cc35169230b04149
excluded: lib/sugar-web/test

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>